Split required stale diagnostic tests
