+++ src/typechecker/tests/resolver_collection/behavior_impls/required_diagnostics.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+mod stale_requires;
+
 #[test]
 fn collect_declarations_with_symbols_reports_resolver_restored_required_target_and_name() {
     let mut program = parse_program(
@@ -159,119 +161,4 @@ Point.requires(Json<i32>)
         "resolver-restored impl refs should keep both required specializations available: {:?}",
         tc.behavior_impls
     );
-}
-
-#[test]
-fn collect_declarations_with_symbols_does_not_fallback_to_stale_ast_behavior_required_metadata() {
-    let mut program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point: { x: i32 }
-
-Point.implements(Json<StaticString>) {
-    encode = (value: Point) StaticString { "point" }
-}
-
-Point.requires(Json<StaticString>)
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_behavior_required_refs_for_test(Namespace::Type, "Point", None);
-    if let Declaration::Requires {
-        behavior_type_args, ..
-    } = &mut program.declarations[3]
-    {
-        behavior_type_args[0] = AstType::I32;
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-backed collection should not validate stale AST-only requires refs when resolver required metadata is incomplete: {:?}",
-            tc.diagnostics
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_does_not_validate_stale_requires_after_target_restore() {
-    let mut program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point: { x: i32 }
-
-Point.implements(Json<StaticString>) {
-    encode = (value: Point) StaticString { "point" }
-}
-
-Point.requires(Json<StaticString>)
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_behavior_required_refs_for_test(Namespace::Type, "Point", None);
-    if let Declaration::Requires {
-        type_name,
-        behavior,
-        behavior_type_args,
-        ..
-    } = &mut program.declarations[3]
-    {
-        *type_name = "Missing".to_string();
-        *behavior = "AlsoMissing".to_string();
-        behavior_type_args[0] = AstType::I32;
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-backed collection should not validate stale AST-only requires refs after target restoration when resolver required metadata is incomplete: {:?}",
-            tc.diagnostics
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_required_name_metadata() {
-    let mut program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point: { x: i32 }
-
-Point.implements(Json<StaticString>) {
-    encode = (value: Point) StaticString { "point" }
-}
-
-Point.requires(Json<StaticString>)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Requires { behavior, .. } = &mut program.declarations[3] {
-        *behavior = "Missing".to_string();
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-restored requires name metadata should avoid stale AST requires diagnostics: {:?}",
-            tc.diagnostics
-        );
 }
